Reload Stations
There are now a new actor : a reload station to reload the base weapon.
There are a structure somewhere to edit easily the parameters of the weapon and of the station.

Bonus : the amount of ammunition is represented by the color of the area of reload (blue for full, red for empty).

Enjoy this update!